refactor(core): decouple client into configuration, routing, and endpoints
