+++ mailgun/_version.py
@@ -1,1 +1,1 @@
-__version__ = "1.7.1"
+__version__ = "1.7.1.post1.dev1"
