+++ mailgun/client.py
@@ -16,29 +16,20 @@ Classes:
 
 from __future__ import annotations
 
-import json
-import logging
-import math
-import re
 import ssl
 import sys
 import warnings
-from enum import Enum
-from functools import lru_cache
-from types import MappingProxyType
-from typing import TYPE_CHECKING, Any, Final, TypeAlias
-from urllib.parse import unquote, urlparse
+from typing import TYPE_CHECKING, Any, Final
 
 import httpx
 import requests  # pyright: ignore[reportMissingModuleSource]
-from requests.adapters import HTTPAdapter  # pyright: ignore[reportMissingModuleSource]
-from requests.exceptions import (
-    ConnectionError as RequestsConnectionError,  # pyright: ignore[reportMissingModuleSource]
-)
 from urllib3.util.retry import Retry
 
-from mailgun import routes
-from mailgun.handlers.error_handler import ApiError, MailgunTimeoutError
+from mailgun.config import Config
+from mailgun.endpoints import AsyncEndpoint, BaseEndpoint, Endpoint
+from mailgun.filters import RedactingFilter
+from mailgun.logger import get_logger
+from mailgun.security import SecretAuth, SecureHTTPAdapter, SecurityGuard
 
 
 if sys.version_info >= (3, 11):
@@ -46,19 +37,9 @@ if sys.version_info >= (3, 11):
 else:
     from typing_extensions import Self
 
-try:
-    from mailgun._version import __version__
-except ImportError:
-    __version__ = "0.0.0-unknown"
-
 
 if TYPE_CHECKING:
     import types
-    from collections.abc import Callable, Mapping
-
-    from httpx import Response as HttpxResponse
-    from requests.models import Response  # pyright: ignore[reportMissingModuleSource]
-
 
 # ==============================================================================
 # 1. PUBLIC API & GLOBALS
@@ -68,686 +49,28 @@ __all__ = [
     "AsyncClient",
     "AsyncEndpoint",
     "BaseClient",
+    "BaseEndpoint",
     "Client",
+    "Config",
     "Endpoint",
+    "RedactingFilter",
+    "SecretAuth",
+    "SecureHTTPAdapter",
+    "SecurityGuard",
 ]
 
-logger = logging.getLogger("mailgun.client")
-if not logger.hasHandlers():
-    logger.addHandler(logging.NullHandler())
+logger = get_logger(__name__)
 
-
-class RedactingFilter(logging.Filter):
-    """Centralized Log Sanitization Filter (CWE-316, CWE-117).
-
-    Scrubs Mailgun private and public key patterns before emitting to logs.
-    """
-
-    SECRET_PATTERN = re.compile(r"(key-|pubkey-)[\w\-]+")
-
-    def filter(self, record: logging.LogRecord) -> bool:
-        # Redact simple string messages
-        if isinstance(record.msg, str):
-            record.msg = self.SECRET_PATTERN.sub(r"\1[REDACTED]", record.msg)
-
-        # Redact formatting arguments if present
-        if isinstance(record.args, dict):
-            record.args = {
-                k: self.SECRET_PATTERN.sub(r"\1[REDACTED]", str(v)) if isinstance(v, str) else v
-                for k, v in record.args.items()
-            }
-        elif isinstance(record.args, tuple):
-            record.args = tuple(
-                self.SECRET_PATTERN.sub(r"\1[REDACTED]", str(v)) if isinstance(v, str) else v
-                for v in record.args
-            )
-        return True
-
-
-logger.addFilter(RedactingFilter())
 
 # Constants for API error handling and logging (fixes Ruff PLR2004)
-_HTTP_ERROR_THRESHOLD: Final[int] = 400
 _MAX_LOG_LENGTH: Final[int] = 500
 _AUTH_TUPLE_LEN: Final = 2
 _TIMEOUT_TUPLE_LEN: Final[int] = 2
 _DEFAULT_TIMEOUT = 60.0
 
-# Type Aliases for SDK Signatures
-TimeoutType: TypeAlias = float | tuple[float, float] | None
 
 # ==============================================================================
-# 2. CORE TYPES & SECURITY GUARDRAILS
-# ==============================================================================
-
-
-class APIVersion(str, Enum):
-    """Constants for Mailgun API versions."""
-
-    V1 = "v1"
-    V2 = "v2"
-    V3 = "v3"
-    V4 = "v4"
-    V5 = "v5"
-
-
-class SecureHTTPAdapter(HTTPAdapter):
-    """Enforce Minimum TLS 1.2+ Protocol Context (MITM & Downgrade Prevention).
-
-    Mitigates CWE-319.
-    """
-
-    def init_poolmanager(self, *args: Any, **kwargs: Any) -> None:
-        context = ssl.create_default_context()
-        context.minimum_version = ssl.TLSVersion.TLSv1_2
-        kwargs["ssl_context"] = context
-        # HTTPAdapter lacks strict static types for this internal method.
-        return super().init_poolmanager(*args, **kwargs)  # type: ignore[no-untyped-call]
-
-
-class SecretAuth(tuple):
-    """OWASP: Obfuscate credentials in memory dumps and tracebacks."""
-
-    __slots__ = ()  # DX & Performance: Prevent __dict__ creation to optimize memory usage.
-
-    def __repr__(self) -> str:
-        return "('api', '***REDACTED***')"
-
-
-class SecurityGuard:
-    """Centralized security validation and sanitization (Defense in Depth).
-
-    This class isolates all Zero-Trust guardrails, enforcing SRP and making it
-    easy to extract into a dedicated security module in future releases.
-    """
-
-    ALLOWED_HTTP_METHODS: Final[frozenset[str]] = frozenset(
-        {"GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS", "HEAD"}
-    )
-    ALLOWED_API_HOSTS: Final[tuple[str, ...]] = (
-        "mailgun.net",
-        "mailgun.org",
-        "localhost",
-        "127.0.0.1",
-    )
-    ALLOWED_KWARGS: Final[frozenset[str]] = frozenset({"proxies", "cert"})
-    SAFE_KEY_PATTERN: Final[re.Pattern[str]] = re.compile(r"^[a-z0-9_]+$")
-    CRLF_SLASH_PATTERN: Final[re.Pattern[str]] = re.compile(r"[\r\n/\\]+")
-
-    @classmethod
-    def sanitize_api_url(cls, raw_url: str) -> str:
-        """Sanitize and validate the base API URL to prevent SSRF and Cleartext transmission.
-
-        Args:
-            raw_url: The raw URL string to sanitize.
-
-        Returns:
-            The sanitized URL string without a trailing slash.
-
-        Raises:
-            ValueError: If the URL uses prohibited cleartext HTTP (CWE-319).
-        """
-        raw_url = raw_url.strip().replace("\r", "").replace("\n", "")
-        parsed = urlparse(raw_url)
-
-        if not parsed.scheme:
-            raw_url = f"https://{raw_url}"
-            parsed = urlparse(raw_url)
-
-        if parsed.scheme == "http" and parsed.hostname not in {"localhost", "127.0.0.1"}:
-            msg = (
-                "CRITICAL SECURITY: Cleartext HTTP transmission is prohibited (CWE-319). Use HTTPS."
-            )
-            raise ValueError(msg)  # Fail Closed
-
-        hostname = parsed.hostname or ""
-        is_valid_host = any(
-            hostname == allowed or hostname.endswith(f".{allowed}")
-            for allowed in cls.ALLOWED_API_HOSTS
-        )
-        if not is_valid_host:
-            msg = (
-                f"SECURITY WARNING: Invalid API host '{hostname}'. Ensure this is a trusted proxy."
-            )
-            logger.warning(msg)
-
-        return raw_url.rstrip("/")
-
-    @classmethod
-    def validate_auth(cls, auth: tuple[str, str] | None) -> tuple[str, str] | None:
-        """Sanitize and validate credentials against Header Injection vulnerabilities.
-
-        Args:
-            auth: A tuple containing the API user and API key, or None.
-
-        Returns:
-            A SecretAuth tuple with cleaned credentials, or None if no auth was provided.
-
-        Raises:
-            ValueError: If the API key contains invalid characters (e.g., newlines).
-        """
-        if auth and isinstance(auth, tuple) and len(auth) == _AUTH_TUPLE_LEN:
-            clean_user = str(auth[0]).strip()
-            clean_key = str(auth[1]).strip()
-
-            if "\n" in clean_key or "\r" in clean_key:
-                raise ValueError("API Key contains invalid characters (Header Injection risk).")
-
-            return SecretAuth((clean_user, clean_key))
-        return auth
-
-    @classmethod
-    def sanitize_key(cls, key: str) -> str:
-        """Normalize and validate the endpoint key from IDE Introspection.
-
-        Args:
-            key: The raw endpoint key to sanitize.
-
-        Returns:
-            The sanitized and validated endpoint key.
-
-        Raises:
-            KeyError: If the resulting key is invalid or empty.
-        """
-        clean_key: str = key.lower()
-        if not cls.SAFE_KEY_PATTERN.fullmatch(clean_key):
-            clean_key = re.sub(r"[^a-z0-9_]", "", clean_key)
-        if not clean_key:
-            msg = f"Invalid endpoint key: {key}"
-            raise KeyError(msg)
-        return clean_key
-
-    @classmethod
-    def sanitize_domain(cls, domain: str | None) -> str | None:
-        """Protect against Path Traversal in URL construction.
-
-        Args:
-            domain: Target domain name to sanitize.
-
-        Returns:
-            The sanitized domain name or None.
-
-        Raises:
-            ValueError: If path traversal characters are detected.
-        """
-        if not domain:
-            return None
-
-        decoded_domain = unquote(domain)
-
-        # Poka-yoke: Actively strip all slashes and newlines (Advanced Traversal & CRLF)
-        safe_domain = cls.CRLF_SLASH_PATTERN.sub("", decoded_domain).strip()
-
-        if ".." in safe_domain:
-            raise ValueError(
-                "CRITICAL SECURITY: Path traversal characters detected in domain parameter."
-            )
-        return safe_domain
-
-    @classmethod
-    def sanitize_http_method(cls, method: str) -> str:
-        """Prevent HTTP Verb Tampering and Attribute Injection.
-
-        Args:
-            method: The HTTP method requested.
-
-        Returns:
-            A safely formatted HTTP method string.
-
-        Raises:
-            ValueError: If the method is not in the allowed list.
-        """
-        safe_method = str(method).strip().upper()
-        if safe_method not in cls.ALLOWED_HTTP_METHODS:
-            msg = f"CRITICAL SECURITY: HTTP method '{safe_method}' is prohibited."
-            raise ValueError(msg)
-        return safe_method
-
-    @classmethod
-    def sanitize_timeout(cls, timeout: TimeoutType) -> TimeoutType:
-        """Prevent Infinite Timeout Thread Exhaustion (DoS).
-
-        Strict Creation-Time Timeout Constraints & Float Validation.
-        Prevents thread pool exhaustion from infinite blocking (CWE-400).
-
-        Args:
-            timeout: The requested timeout value.
-
-        Returns:
-            The safely verified timeout value.
-
-        Raises:
-            ValueError: If the timeout is a negative number, zero, non-finite,
-                or a tuple with an incorrect number of elements.
-        """
-        if timeout is None:
-            # Soft Deprecation
-            warnings.warn(
-                "Passing 'timeout=None' allows infinite socket blocking (CWE-400). "
-                "This will be removed in a future major release. Please provide an explicit timeout.",
-                DeprecationWarning,
-                stacklevel=3,
-            )
-            return None
-
-        def _validate_float(val: Any) -> float:
-            """Validate float value.
-
-            Args:
-                val: The timeout value.
-
-            Returns:
-                The timeout float value.
-
-            Raises:
-                TypeError: If the timeout is not a numeric type.
-                ValueError: If the timeout is NaN, Infinity, or less than or equal to zero.
-            """
-            if isinstance(val, bool) or not isinstance(val, (int, float)):
-                msg = f"Timeout must be a numeric value, got {type(val).__name__}"
-                raise TypeError(msg)
-
-            f_val = float(val)
-
-            if math.isnan(f_val) or math.isinf(f_val):
-                raise ValueError("Timeout must be a finite number.")
-            if f_val <= 0:
-                raise ValueError("Timeout must be a strictly positive finite number.")
-            return f_val
-
-        if isinstance(timeout, tuple):
-            expected_tuple_length = 2
-            if len(timeout) != expected_tuple_length:
-                raise ValueError(
-                    "Timeout must be a tuple containing exactly two elements: (connect, read)."
-                )
-            return (_validate_float(timeout[0]), _validate_float(timeout[1]))
-
-        return _validate_float(timeout)
-
-    @classmethod
-    def filter_safe_kwargs(cls, kwargs: dict[str, Any]) -> dict[str, Any]:
-        """Prevent Mass Assignment of internal HTTP client states.
-
-        Args:
-            kwargs: Dictionary of keyword arguments passed to the network layer.
-
-        Returns:
-            A filtered dictionary containing only allowed low-level HTTP settings.
-        """
-        return {k: v for k, v in kwargs.items() if k in cls.ALLOWED_KWARGS}
-
-    @staticmethod
-    def sanitize_headers(headers: dict[str, str] | None) -> dict[str, str] | None:
-        """Poka-yoke: Prevent HTTP Header Injection (CWE-113).
-
-        Returns:
-            The sanitized headers dictionary, or None if no headers were provided.
-
-        Raises:
-            ValueError: If a CRLF injection pattern is detected in any header key or value.
-        """
-        if not headers:
-            return headers
-        for key, value in headers.items():
-            # Check both key and value
-            if "\n" in str(key) or "\r" in str(key) or "\n" in str(value) or "\r" in str(value):
-                msg = f"CRLF injection detected in header: {key}"
-                raise ValueError(msg)
-        return headers
-
-
-# ==============================================================================
-# 3. ROUTING ENGINE & CONFIGURATION
-# ==============================================================================
-
-
-@lru_cache(maxsize=32)
-def _load_handler(endpoint_key: str) -> Callable[..., str]:  # noqa: PLR0911, PLR0912
-    """Lazy load the API URL handler for a specific endpoint using SAST-safe literal imports.
-
-    This maintains zero-I/O startup performance. The lru_cache ensures this branching logic
-    is executed exactly once per route type.
-
-    Returns:
-        Callable: The specific handler function for the requested endpoint.
-    """
-    # Group 1: Domains Handler (Most common aliases grouped for speed)
-    if endpoint_key in {"domains", "dkim_authority", "dkim_selector", "web_prefix"}:
-        from mailgun.handlers.domains_handler import handle_domains  # noqa: PLC0415
-
-        return handle_domains
-    if endpoint_key == "domainlist":
-        from mailgun.handlers.domains_handler import handle_domainlist  # noqa: PLC0415
-
-        return handle_domainlist
-    if endpoint_key == "dkim":
-        from mailgun.handlers.domains_handler import handle_dkimkeys  # noqa: PLC0415
-
-        return handle_dkimkeys
-    if endpoint_key == "sending_queues":
-        from mailgun.handlers.domains_handler import handle_sending_queues  # noqa: PLC0415
-
-        return handle_sending_queues
-    if endpoint_key == "mailboxes":
-        from mailgun.handlers.domains_handler import handle_mailboxes_credentials  # noqa: PLC0415
-
-        return handle_mailboxes_credentials
-    if endpoint_key == "webhooks":
-        from mailgun.handlers.domains_handler import handle_webhooks  # noqa: PLC0415
-
-        return handle_webhooks
-
-    # Group 2: Suppressions
-    if endpoint_key == "bounces":
-        from mailgun.handlers.suppressions_handler import handle_bounces  # noqa: PLC0415
-
-        return handle_bounces
-    if endpoint_key == "unsubscribes":
-        from mailgun.handlers.suppressions_handler import handle_unsubscribes  # noqa: PLC0415
-
-        return handle_unsubscribes
-    if endpoint_key == "whitelists":
-        from mailgun.handlers.suppressions_handler import handle_whitelists  # noqa: PLC0415
-
-        return handle_whitelists
-    if endpoint_key == "complaints":
-        from mailgun.handlers.suppressions_handler import handle_complaints  # noqa: PLC0415
-
-        return handle_complaints
-
-    # Group 3: Specific Services
-    if endpoint_key == "resendmessage":
-        from mailgun.handlers.messages_handler import handle_resend_message  # noqa: PLC0415
-
-        return handle_resend_message
-    if endpoint_key == "ips":
-        from mailgun.handlers.ips_handler import handle_ips  # noqa: PLC0415
-
-        return handle_ips
-    if endpoint_key == "ip_pools":
-        from mailgun.handlers.ip_pools_handler import handle_ippools  # noqa: PLC0415
-
-        return handle_ippools
-    if endpoint_key == "tags":
-        from mailgun.handlers.tags_handler import handle_tags  # noqa: PLC0415
-
-        return handle_tags
-    if endpoint_key == "routes":
-        from mailgun.handlers.routes_handler import handle_routes  # noqa: PLC0415
-
-        return handle_routes
-    if endpoint_key == "lists":
-        from mailgun.handlers.mailinglists_handler import handle_lists  # noqa: PLC0415
-
-        return handle_lists
-    if endpoint_key == "templates":
-        from mailgun.handlers.templates_handler import handle_templates  # noqa: PLC0415
-
-        return handle_templates
-    if endpoint_key == "addressvalidate":
-        from mailgun.handlers import email_validation_handler as evh  # noqa: PLC0415
-
-        return evh.handle_address_validate
-    if endpoint_key == "inbox":
-        from mailgun.handlers.inbox_placement_handler import handle_inbox  # noqa: PLC0415
-
-        return handle_inbox
-    if endpoint_key == "analytics":
-        from mailgun.handlers.metrics_handler import handle_metrics  # noqa: PLC0415
-
-        return handle_metrics
-    if endpoint_key == "bounce-classification":
-        from mailgun.handlers import bounce_classification_handler as bch  # noqa: PLC0415
-
-        return bch.handle_bounce_classification
-    if endpoint_key == "users":
-        from mailgun.handlers.users_handler import handle_users  # noqa: PLC0415
-
-        return handle_users
-    if endpoint_key == "keys":
-        from mailgun.handlers.keys_handler import handle_keys  # noqa: PLC0415
-
-        return handle_keys
-
-    # Group 4: Fallback for "messages", "messages.mime", "events", and unknown routes
-    from mailgun.handlers.default_handler import handle_default  # noqa: PLC0415
-
-    return handle_default
-
-
-@lru_cache
-def _get_cached_route_data(clean_key: str) -> dict[str, Any]:
-    """Apply internal cached routing logic.
-
-    Uses only hashable types (str) as arguments to avoid TypeError.
-
-    Args:
-        clean_key: The sanitized endpoint key.
-
-    Returns:
-        A dictionary containing versioning and path data for the route.
-    """
-    # Resolve virtual property aliases before processing
-    clean_key = routes.ROUTE_ALIASES.get(clean_key, clean_key)
-
-    if clean_key in routes.EXACT_ROUTES:
-        version, route_keys = routes.EXACT_ROUTES[clean_key]
-        return {"version": version, "keys": tuple(route_keys)}
-
-    route_parts = clean_key.split("_")
-    primary_resource = route_parts[0]
-
-    if primary_resource == "domains":
-        return {"type": "domain", "parts": tuple(route_parts)}
-
-    if primary_resource in routes.PREFIX_ROUTES:
-        version, suffix, key_override = routes.PREFIX_ROUTES[primary_resource]
-        final_parts = route_parts.copy()
-        if key_override:
-            final_parts[0] = key_override
-        return {"version": version, "suffix": suffix, "keys": tuple(final_parts)}
-
-    return {"version": APIVersion.V3.value, "keys": tuple(route_parts)}
-
-
-class Config:
-    """Configuration engine for the Mailgun API client.
-
-    Using a data-driven routing approach.
-    """
-
-    __slots__ = ("_baked_urls", "api_url", "ex_handler")
-
-    DEFAULT_API_URL: Final[str] = "https://api.mailgun.net"
-    USER_AGENT: Final[str] = f"mailgun-api-python/{__version__}"
-
-    # Use Mapping to denote read-only dictionary-like structures
-    _HEADERS_BASE: Final[Mapping[str, str]] = MappingProxyType({"User-agent": USER_AGENT})
-    _HEADERS_JSON: Final[Mapping[str, str]] = MappingProxyType(
-        {"User-agent": USER_AGENT, "Content-Type": "application/json"}
-    )
-
-    # --- ENCAPSULATED ROUTING REGISTRIES ---
-    _DOMAINS_RESOURCE: Final[str] = "domains"
-
-    # Mapping[str, Any] is used because the values in routes vary in structure
-    _EXACT_ROUTES: Final[Mapping[str, Any]] = MappingProxyType(routes.EXACT_ROUTES)
-    _PREFIX_ROUTES: Final[Mapping[str, Any]] = MappingProxyType(routes.PREFIX_ROUTES)
-    _DOMAIN_ALIASES: Final[Mapping[str, str]] = MappingProxyType(routes.DOMAIN_ALIASES)
-
-    _DOMAIN_ENDPOINTS: Final[Mapping[str, tuple[str, ...]]] = MappingProxyType(
-        routes.DOMAIN_ENDPOINTS
-    )
-    _V1_ENDPOINTS: Final[frozenset[str]] = frozenset(routes.DOMAIN_ENDPOINTS["v1"])
-    _V3_ENDPOINTS: Final[frozenset[str]] = frozenset(routes.DOMAIN_ENDPOINTS["v3"])
-    _V4_ENDPOINTS: Final[frozenset[str]] = frozenset(routes.DOMAIN_ENDPOINTS.get("v4", []))
-
-    def __init__(self, api_url: str | None = None) -> None:
-        """Initialize the configuration engine.
-
-        Args:
-            api_url: Optional custom base URL for the Mailgun API.
-        """
-        self.ex_handler: bool = True
-        base_url_input: str = api_url or self.DEFAULT_API_URL
-
-        self.api_url: str = self._normalize_api_url(base_url_input)
-
-        self._baked_urls: Final[dict[str, str]] = {
-            ver.value: f"{self.api_url}/{ver.value}" for ver in APIVersion
-        }
-
-    @staticmethod
-    def _normalize_api_url(raw_url: str) -> str:
-        """Validates and normalizes the base API URL.
-
-        Ensures no explicit versions are embedded in the path that would break
-        dynamic f-string routing.
-
-        Args:
-            raw_url: The raw base URL string provided by the user.
-
-        Returns:
-            The sanitized and normalized API URL string.
-
-        Raises:
-            ApiError: If an ambiguous API version is found embedded within the custom path.
-        """
-        safe_url: str = SecurityGuard.sanitize_api_url(raw_url)
-
-        parsed = urlparse(safe_url)
-        path_segments = [seg for seg in parsed.path.split("/") if seg]
-
-        known_versions = {ver.value for ver in APIVersion}
-
-        # Ambiguity & Backward Compatibility Check
-        for i, segment in enumerate(path_segments):
-            if segment in known_versions:
-                is_last_segment = i == len(path_segments) - 1
-
-                if is_last_segment:
-                    safe_url = safe_url.removesuffix(f"/{segment}")
-                    logger.warning(
-                        "Semantic Configuration Warning: 'api_url' should be the base domain. The trailing '%s' was stripped to prevent routing duplication.",
-                        segment,
-                    )
-                else:
-                    # Fail-Fast: The version is trapped inside a complex path
-                    msg = (
-                        f"Ambiguous API URL configuration: '{raw_url}'.\n"
-                        f"The SDK automatically handles version routing, but an explicit "
-                        f"version ('{segment}') was found embedded within your custom path. "
-                        f"Please provide only the base host (e.g., 'https://api.mailgun.net')."
-                    )
-                    raise ApiError(msg)
-
-        return safe_url
-
-    def _build_base_url(self, version: APIVersion | str, suffix: str = "") -> str:
-        """Construct API URL with precise slash control to prevent 404s.
-
-        Args:
-            version: The API version to use.
-            suffix: An optional suffix to append to the base URL.
-
-        Returns:
-            The fully constructed base URL string.
-        """
-        ver_str: str = version.value if isinstance(version, APIVersion) else version
-        # O(1) access instead of dynamic concatenation, ensuring no trailing slash
-        base: str = self._baked_urls.get(ver_str, f"{self.api_url}/{ver_str}").rstrip("/")
-
-        if suffix:
-            path: str = f"{suffix}/" if suffix == self._DOMAINS_RESOURCE else suffix
-            return f"{base}/{path}"
-
-        return f"{base}/"
-
-    def _resolve_domains_route(self, route_parts: list[str]) -> dict[str, Any]:
-        """Handle context-aware versioning for domain-related endpoints.
-
-        Args:
-            route_parts: The components of the route requested.
-
-        Returns:
-            A dictionary containing a string base URL and a tuple of keys.
-        """
-        if any(action in route_parts for action in ("activate", "deactivate")):
-            return {
-                "base": self._build_base_url(APIVersion.V4),
-                "keys": (
-                    self._DOMAINS_RESOURCE,
-                    "{authority_name}",
-                    "keys",
-                    "{selector}",
-                    route_parts[-1],
-                ),
-            }
-
-        mapped_parts: list[str] = [self._DOMAIN_ALIASES.get(p, p) for p in route_parts]
-
-        if not mapped_parts or mapped_parts[0] != self._DOMAINS_RESOURCE:
-            mapped_parts.insert(0, self._DOMAINS_RESOURCE)
-
-        version: APIVersion = APIVersion.V3
-
-        if len(mapped_parts) > 1:
-            for part in reversed(mapped_parts[1:]):
-                if part in self._V1_ENDPOINTS:
-                    version = APIVersion.V1
-                    break
-                if part in self._V4_ENDPOINTS:
-                    version = APIVersion.V4
-                    break
-                if part in self._V3_ENDPOINTS:
-                    version = APIVersion.V3
-                    break
-
-        return {
-            "base": self._build_base_url(version, self._DOMAINS_RESOURCE),
-            "keys": mapped_parts.copy(),
-        }
-
-    def __getitem__(self, key: str) -> tuple[dict[str, Any], dict[str, str]]:
-        """Retrieve the URL configuration and headers for a specific endpoint.
-
-        Args:
-            key: The name of the endpoint route (e.g., 'messages', 'bounces').
-
-        Returns:
-            A tuple containing the URL configuration dictionary and the headers dictionary.
-        """
-        clean_key = SecurityGuard.sanitize_key(key)
-
-        route_data = _get_cached_route_data(clean_key)
-
-        # HTTP header mapping based on endpoint naming conventions
-        requires_json_headers = "analytics" in clean_key or "bounceclassification" in clean_key
-
-        # Prepare headers
-        headers_map = self._HEADERS_JSON if requires_json_headers else self._HEADERS_BASE
-        headers = dict(headers_map)
-
-        # Reconstruct result
-        if route_data.get("type") == "domain":
-            return self._resolve_domains_route(list(route_data["parts"])), headers
-
-        safe_url = {
-            "base": self._build_base_url(route_data["version"], route_data.get("suffix", "")),
-            "keys": list(route_data["keys"]),
-        }
-
-        return safe_url, headers
-
-    @property
-    def available_endpoints(self) -> set[str]:
-        """Provide public access to valid route keys for IDE introspection."""
-        return set(self._EXACT_ROUTES.keys()) | set(self._PREFIX_ROUTES.keys())
-
-
-# ==============================================================================
-# 4. BASE CLASSES (Abstract Interfaces)
+# 1. BASE CLASS (Abstract Interface)
 # ==============================================================================
 
 
@@ -813,148 +136,8 @@ class BaseClient:
         return sorted(set(super().__dir__()) | self.config.available_endpoints)
 
 
-class BaseEndpoint:
-    """Base class for endpoints. Contains methods common for Endpoint and AsyncEndpoint."""
-
-    __slots__ = ("_auth", "_timeout", "_url", "headers")
-
-    def __init__(
-        self,
-        url: dict[str, Any],
-        headers: dict[str, str],
-        auth: tuple[str, str] | None,
-        timeout: TimeoutType = 60,
-    ) -> None:
-        """Initialize a new BaseEndpoint instance.
-
-        Args:
-            url: URL dictionary with pairs {"base": "keys"}.
-            headers: Headers dictionary.
-            auth: Authentication tuple or None.
-        """
-        self._url = url
-        self.headers = headers
-        self._auth = auth
-        self._timeout = timeout
-
-    @staticmethod
-    def _warn_if_deprecated(method: str, target_url: str) -> None:
-        """Check the formulated URL against the registry of deprecated endpoints.
-
-        Issues both a standard Python DeprecationWarning and an SDK logger warning.
-
-        Args:
-            method: Requested HTTP method.
-            target_url: Formulated destination URL.
-        """
-        path = urlparse(target_url).path
-
-        # Iterate over the dynamically compiled, cached regexes
-        for pattern, msg in routes.get_deprecated_regexes().items():
-            if pattern.search(path):
-                warning_message = f"DEPRECATED API CALL ({method.upper()} {path}): {msg}"
-                warnings.warn(warning_message, DeprecationWarning, stacklevel=3)
-                logger.warning(warning_message)
-                break
-
-    def __repr__(self) -> str:
-        """DX: Show the actual resolved target route instead of memory address.
-
-        Returns:
-            A string representation of the Endpoint and its target route.
-        """
-        route_path = "/".join(self._url.get("keys", ["unknown"]))
-        return f"<{self.__class__.__name__} target='/{route_path}'>"
-
-    @staticmethod
-    def build_url(
-        url: dict[str, Any],
-        domain: str | None = None,
-        method: str | None = None,
-        **kwargs: Any,
-    ) -> str:
-        """Build the final request URL using predefined handlers.
-
-        Note: Some URLs are built in the Config class as they cannot be generated dynamically.
-
-        Args:
-            url: Incoming URL structure containing base and keys.
-            domain: Target domain name.
-            method: Requested HTTP method.
-            **kwargs: Additional arguments required by specific handlers.
-
-        Returns:
-            The fully constructed target URL.
-
-        Raises:
-            ApiError: If the domain is required but missing.
-        """
-        keys = url.get("keys", [])
-        endpoint_key = keys[0] if keys else ""
-
-        if not domain and endpoint_key == "messages":
-            raise ApiError("Domain is required")
-
-        # Load the handler function dynamically via the cached lazy loader
-        handler = _load_handler(endpoint_key)
-
-        return handler(url, domain, method, **kwargs)  # type: ignore[no-untyped-call]
-
-    def _merge_headers(self, kwargs: dict[str, Any]) -> dict[str, str]:
-        """Safely extract and merge custom headers from kwargs.
-
-        Returns:
-            A dictionary containing the safely merged headers.
-        """
-        custom_headers = kwargs.pop("headers", {})
-        req_headers = self.headers.copy()
-
-        if custom_headers and isinstance(custom_headers, dict):
-            req_headers.update(custom_headers)
-
-        return req_headers
-
-    def _prepare_request(
-        self,
-        method: str,
-        url: dict[str, Any],
-        domain: str | None,
-        timeout: TimeoutType,
-        headers: dict[str, str],
-        kwargs: dict[str, Any],
-    ) -> tuple[str, str, str, TimeoutType, dict[str, str], dict[str, Any]]:
-        """Security and routing preparation logic.
-
-        Args:
-            method: The requested HTTP method.
-            url: Incoming URL structure containing base and keys.
-            domain: Target domain name to sanitize.
-            timeout: Request timeout duration.
-            headers: Headers dictionary.
-            kwargs: Additional keyword arguments.
-
-        Returns:
-            A tuple containing safe_method, target_url, safe_url_for_log, safe_timeout, safe_headers, and safe_kwargs.
-        """
-        safe_method = SecurityGuard.sanitize_http_method(method)
-        safe_kwargs = SecurityGuard.filter_safe_kwargs(kwargs)
-        safe_headers = SecurityGuard.sanitize_headers(headers) or {}
-        target_domain = SecurityGuard.sanitize_domain(domain)
-
-        actual_timeout = timeout if timeout is not None else self._timeout
-        safe_timeout = SecurityGuard.sanitize_timeout(actual_timeout)
-
-        target_url = self.build_url(url, domain=target_domain, method=safe_method, **kwargs)
-        self._warn_if_deprecated(safe_method, target_url)
-
-        # PEP 578 and protection against Log Forging (CWE-117)
-        safe_url_for_log = target_url.replace("\n", "_").replace("\r", "_")
-
-        return safe_method, target_url, safe_url_for_log, safe_timeout, safe_headers, safe_kwargs
-
-
 # ==============================================================================
-# 5. SYNCHRONOUS IMPLEMENTATION
+# 2. SYNCHRONOUS IMPLEMENTATION
 # ==============================================================================
 
 
@@ -1036,9 +219,14 @@ class Client(BaseClient):
 
     def close(self) -> None:
         """Close the underlying requests.Session connection pool and purge memory."""
-        self._session.auth = None
-        self._session.headers.clear()
-        self._session.close()
+        if self._session:
+            try:
+                # CWE-316: Clear session resources
+                self._session.auth = None
+                self._session.headers.clear()
+                self._session.close()
+            finally:
+                self._session = None
         self.auth = None
 
     def __enter__(self) -> Self:
@@ -1059,564 +247,9 @@ class Client(BaseClient):
         self.close()
 
 
-class Endpoint(BaseEndpoint):
-    """Generate synchronous requests and return responses."""
-
-    __slots__ = ("_session",)
-
-    def __init__(
-        self,
-        url: dict[str, Any],
-        headers: dict[str, str],
-        auth: tuple[str, str] | None = None,
-        session: requests.Session | None = None,
-        timeout: TimeoutType = 60,
-    ) -> None:
-        """Initialize a new Endpoint instance for synchronous API interaction.
-
-        Args:
-            url: URL dictionary with pairs {"base": "keys"}.
-            headers: Headers dictionary.
-            auth: requests auth tuple or None.
-            session: Optional pre-configured requests.Session instance.
-        """
-        super().__init__(url, headers, auth, timeout=timeout)
-        self._session = session or requests.Session()
-
-    def api_call(
-        self,
-        auth: tuple[str, str] | None,
-        method: str,
-        url: dict[str, Any],
-        headers: dict[str, str],
-        data: Any | None = None,
-        filters: Mapping[str, str | Any] | None = None,
-        timeout: TimeoutType = None,
-        files: Any | None = None,
-        domain: str | None = None,
-        **kwargs: Any,
-    ) -> Response | Any:
-        """Execute the HTTP request to the Mailgun API.
-
-        Args:
-            auth: Authentication tuple.
-            method: The HTTP method to use (e.g., 'GET', 'POST', 'PUT', 'DELETE').
-            url: The final formulated endpoint URL dictionary.
-            headers: Request headers.
-            data: Payload data (form data or JSON).
-            filters: Query parameters.
-            timeout: Request timeout duration in seconds.
-            files: Files to upload.
-            domain: Target domain name.
-            **kwargs: Additional parameters to be passed to the underlying HTTP client.
-
-        Returns:
-            The HTTP response object from the server.
-
-        Raises:
-            MailgunTimeoutError: If the request times out.
-            ApiError: If the server returns a 4xx or 5xx status code or a network error occurs.
-        """
-        safe_method, target_url, safe_url_for_log, safe_timeout, safe_headers, safe_kwargs = (
-            self._prepare_request(method, url, domain, timeout, headers, kwargs)
-        )
-
-        # Case-insensitive validation for Content-Type to conform with RFC 7230
-        is_json_request = any(
-            k.lower() == "content-type" and "application/json" in str(v).lower()
-            for k, v in safe_headers.items()
-        )
-
-        if is_json_request and data is not None and not isinstance(data, (str, bytes)):
-            data = json.dumps(data, separators=(",", ":"))
-
-        req_method = getattr(self._session, safe_method.lower())
-
-        sys.audit("mailgun.api.request", safe_method.upper(), safe_url_for_log)
-        logger.debug("Sending Request: %s %s", safe_method.upper(), safe_url_for_log)
-
-        try:
-            response = req_method(
-                target_url,
-                data=data,
-                params=filters,
-                headers=safe_headers,
-                auth=auth,
-                timeout=safe_timeout,
-                files=files,
-                verify=True,
-                stream=False,
-                allow_redirects=False,
-                **safe_kwargs,
-            )
-
-            status_code = getattr(response, "status_code", 200)
-            is_error = isinstance(status_code, int) and status_code >= _HTTP_ERROR_THRESHOLD
-            if is_error:
-                logger.error(
-                    "API Error %s | %s %s", status_code, safe_method.upper(), safe_url_for_log
-                )
-            else:
-                logger.debug(
-                    "API Success %s | %s %s",
-                    getattr(response, "status_code", 200),
-                    safe_method.upper(),
-                    target_url,
-                )
-
-        except requests.exceptions.Timeout as e:
-            logger.exception("Timeout Error: %s %s", safe_method.upper(), safe_url_for_log)
-            raise MailgunTimeoutError("Request timed out") from e
-        except RequestsConnectionError as e:
-            logger.critical("Connection Failed (DNS/Network): %s | URL: %s", e, safe_url_for_log)
-            msg = f"Network routing failed: {e}"
-            raise ApiError(msg) from e
-        except requests.RequestException as e:
-            logger.critical("Request Exception: %s | URL: %s", e, safe_url_for_log)
-            raise ApiError(e) from e
-        else:
-            return response
-
-    def get(
-        self,
-        filters: Mapping[str, str | Any] | None = None,
-        domain: str | None = None,
-        **kwargs: Any,
-    ) -> Response:
-        """Send a GET request to retrieve resources.
-
-        Args:
-            filters: Query parameters to include in the request.
-            domain: Target domain name.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return self.api_call(
-            self._auth,
-            "get",
-            self._url,
-            domain=domain,
-            headers=merged_headers,
-            filters=filters,
-            **kwargs,
-        )
-
-    def create(
-        self,
-        data: Any | None = None,
-        filters: Mapping[str, str | Any] | None = None,
-        domain: str | None = None,
-        headers: Any = None,
-        files: Any | None = None,
-        **kwargs: Any,
-    ) -> Response:
-        """Send a POST request to create a new resource or execute an action.
-
-        Args:
-            data: Payload data (form data or JSON) to include in the request.
-            filters: Query parameters to include in the request.
-            domain: Target domain name.
-            headers: Additional headers to merge with the default headers.
-            files: Files to upload in the request.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        if headers is not None:
-            kwargs["headers"] = headers
-        merged_headers = self._merge_headers(kwargs)
-
-        return self.api_call(
-            self._auth,
-            "post",
-            self._url,
-            files=files,
-            domain=domain,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    def put(
-        self, data: Any | None = None, filters: Mapping[str, str | Any] | None = None, **kwargs: Any
-    ) -> Response:
-        """Send a PUT request to update or replace a resource.
-
-        Args:
-            data: Payload data to include in the request.
-            filters: Query parameters to include in the request.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return self.api_call(
-            self._auth,
-            "put",
-            self._url,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    def patch(
-        self, data: Any | None = None, filters: Mapping[str, str | Any] | None = None, **kwargs: Any
-    ) -> Response:
-        """Send a PATCH request to partially update a resource.
-
-        Args:
-            data: Payload data to include in the request.
-            filters: Query parameters to include in the request.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return self.api_call(
-            self._auth,
-            "patch",
-            self._url,
-            data=data,
-            headers=merged_headers,
-            filters=filters,
-            **kwargs,
-        )
-
-    def update(
-        self, data: Any | None, filters: Mapping[str, str | Any] | None = None, **kwargs: Any
-    ) -> Response:
-        """Send a PUT request specifically structured for updating resources with dynamic headers.
-
-        Args:
-            data: Payload data (form data or JSON).
-            filters: Query parameters to include in the request.
-            **kwargs: Additional arguments, including custom 'headers', to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return self.api_call(
-            self._auth,
-            "put",
-            self._url,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    def delete(self, domain: str | None = None, **kwargs: Any) -> Response:
-        """Send a DELETE request to remove a resource.
-
-        Args:
-            domain: Target domain name.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return self.api_call(
-            self._auth, "delete", self._url, headers=merged_headers, domain=domain, **kwargs
-        )
-
-
 # ==============================================================================
-# 6. ASYNCHRONOUS IMPLEMENTATION
+# 2. ASYNC IMPLEMENTATION
 # ==============================================================================
-
-
-class AsyncEndpoint(BaseEndpoint):
-    """Generate async requests and return responses using httpx."""
-
-    __slots__ = ("_client",)
-
-    def __init__(
-        self,
-        url: dict[str, Any],
-        headers: dict[str, str],
-        auth: tuple[str, str] | None,
-        client: httpx.AsyncClient | None = None,
-        timeout: TimeoutType = 60,
-    ) -> None:
-        """Initialize a new AsyncEndpoint instance for asynchronous API interaction.
-
-        Args:
-            url: URL dictionary with pairs {"base": "keys"}.
-            headers: Headers dictionary.
-            auth: httpx auth tuple or None.
-            client: Optional httpx.AsyncClient instance to reuse.
-        """
-        super().__init__(url, headers, auth, timeout=timeout)
-        self._client = client or httpx.AsyncClient()
-
-    async def api_call(
-        self,
-        auth: tuple[str, str] | None,
-        method: str,
-        url: dict[str, Any],
-        headers: dict[str, str],
-        data: Any | None = None,
-        filters: Mapping[str, str | Any] | None = None,
-        timeout: TimeoutType = None,
-        files: Any | None = None,
-        domain: str | None = None,
-        **kwargs: Any,
-    ) -> HttpxResponse:
-        """Execute the asynchronous HTTP request to the Mailgun API.
-
-        Args:
-            auth: Authentication tuple.
-            method: The HTTP method to use (e.g., 'GET', 'POST', 'PUT', 'DELETE').
-            url: The final formulated endpoint URL dictionary.
-            headers: Request headers.
-            data: Payload data (form data or JSON).
-            filters: Query parameters.
-            timeout: Request timeout duration in seconds.
-            files: Files to upload.
-            domain: Target domain name.
-            **kwargs: Additional parameters to be passed to the underlying HTTP client.
-
-        Returns:
-            The HTTP response object from the server.
-
-        Raises:
-            MailgunTimeoutError: If the request times out.
-            ApiError: If the server returns a 4xx or 5xx status code or a network error occurs.
-        """
-        safe_method, target_url, safe_url_for_log, safe_timeout, safe_headers, safe_kwargs = (
-            self._prepare_request(method, url, domain, timeout, headers, kwargs)
-        )
-
-        if isinstance(safe_timeout, tuple):
-            safe_timeout = httpx.Timeout(safe_timeout[1], connect=safe_timeout[0])
-
-        # Case-insensitive validation for Content-Type to conform with RFC 7230
-        is_json_request = any(
-            k.lower() == "content-type" and "application/json" in str(v).lower()
-            for k, v in safe_headers.items()
-        )
-
-        if is_json_request and data is not None and not isinstance(data, (str, bytes)):
-            data = json.dumps(data, separators=(",", ":"))
-
-        request_kwargs: dict[str, Any] = {
-            "method": safe_method.upper(),
-            "url": target_url,
-            "params": filters,
-            "files": files,
-            "headers": safe_headers,
-            "auth": auth,
-            "timeout": safe_timeout,
-            "follow_redirects": False,
-        }
-
-        # Safe kwargs passthrough (e.g., allow_redirects)
-        request_kwargs.update(safe_kwargs)
-
-        if isinstance(data, (str, bytes)):
-            request_kwargs["content"] = data
-        else:
-            request_kwargs["data"] = data
-
-        # PEP 578 and protection against Log Forging (CWE-117)
-        safe_url_for_log = target_url.replace("\n", "_").replace("\r", "_")
-        sys.audit("mailgun.api.request", safe_method.upper(), safe_url_for_log)
-        logger.debug("Sending Async Request: %s %s", safe_method.upper(), safe_url_for_log)
-
-        try:
-            response = await self._client.request(**request_kwargs)
-
-            status_code = getattr(response, "status_code", 200)
-            is_error = isinstance(status_code, int) and status_code >= _HTTP_ERROR_THRESHOLD
-            if is_error:
-                logger.error(
-                    "API Error %s | %s %s", status_code, safe_method.upper(), safe_url_for_log
-                )
-            else:
-                logger.debug(
-                    "API Success %s | %s %s",
-                    getattr(response, "status_code", 200),
-                    safe_method.upper(),
-                    target_url,
-                )
-
-        except httpx.TimeoutException as e:
-            logger.exception("Timeout Error: %s %s", safe_method.upper(), safe_url_for_log)
-            raise MailgunTimeoutError("Request timed out") from e
-        except httpx.ConnectError as e:
-            logger.critical(
-                "Async Connection Failed (DNS/Network): %s | URL: %s", e, safe_url_for_log
-            )
-            msg = f"Network routing failed: {e}"
-            raise ApiError(msg) from e
-        except httpx.RequestError as e:
-            logger.critical("Request Exception: %s | URL: %s", e, safe_url_for_log)
-            raise ApiError(e) from e
-        else:
-            return response
-
-    async def get(
-        self,
-        filters: Mapping[str, str | Any] | None = None,
-        domain: str | None = None,
-        **kwargs: Any,
-    ) -> HttpxResponse:
-        """Send an asynchronous GET request to retrieve resources.
-
-        Args:
-            filters: Query parameters to include in the request.
-            domain: Target domain name.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return await self.api_call(
-            self._auth,
-            "get",
-            self._url,
-            domain=domain,
-            headers=merged_headers,
-            filters=filters,
-            **kwargs,
-        )
-
-    async def create(
-        self,
-        data: Any | None = None,
-        filters: Mapping[str, str | Any] | None = None,
-        domain: str | None = None,
-        headers: Any = None,
-        files: Any | None = None,
-        **kwargs: Any,
-    ) -> HttpxResponse:
-        """Send an asynchronous POST request to create a new resource or execute an action.
-
-        Args:
-            data: Payload data (form data or JSON) to include in the request.
-            filters: Query parameters to include in the request.
-            domain: Target domain name.
-            headers: Additional headers to merge with the default headers.
-            files: Files to upload in the request.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        if headers is not None:
-            kwargs["headers"] = headers
-        merged_headers = self._merge_headers(kwargs)
-
-        return await self.api_call(
-            self._auth,
-            "post",
-            self._url,
-            files=files,
-            domain=domain,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    async def put(
-        self, data: Any | None = None, filters: Mapping[str, str | Any] | None = None, **kwargs: Any
-    ) -> HttpxResponse:
-        """Send an asynchronous PUT request to update or replace a resource.
-
-        Args:
-            data: Payload data to include in the request.
-            filters: Query parameters to include in the request.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return await self.api_call(
-            self._auth,
-            "put",
-            self._url,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    async def patch(
-        self, data: Any | None = None, filters: Mapping[str, str | Any] | None = None, **kwargs: Any
-    ) -> HttpxResponse:
-        """Send an asynchronous PATCH request to partially update a resource.
-
-        Args:
-            data: Payload data to include in the request.
-            filters: Query parameters to include in the request.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return await self.api_call(
-            self._auth,
-            "patch",
-            self._url,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    async def update(
-        self, data: Any | None, filters: Mapping[str, str | Any] | None = None, **kwargs: Any
-    ) -> HttpxResponse:
-        """Send an asynchronous PUT request specifically structured for updating resources with dynamic headers.
-
-        Args:
-            data: Payload data (form data or JSON).
-            filters: Query parameters to include in the request.
-            **kwargs: Additional arguments, including custom 'headers', to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-
-        return await self.api_call(
-            self._auth,
-            "put",
-            self._url,
-            headers=merged_headers,
-            data=data,
-            filters=filters,
-            **kwargs,
-        )
-
-    async def delete(self, domain: str | None = None, **kwargs: Any) -> httpx.Response:
-        """Send an asynchronous DELETE request to remove a resource.
-
-        Args:
-            domain: Target domain name.
-            **kwargs: Additional arguments to pass to the HTTP client.
-
-        Returns:
-            The HTTP response object.
-        """
-        merged_headers = self._merge_headers(kwargs)
-        return await self.api_call(
-            self._auth, "delete", self._url, headers=merged_headers, domain=domain, **kwargs
-        )
 
 
 class AsyncClient(BaseClient):
@@ -1673,7 +306,7 @@ class AsyncClient(BaseClient):
             raise AttributeError(msg) from None
 
     @property
-    def _client(self) -> httpx.AsyncClient:
+    def _client(self) -> httpx.AsyncClient | None:
         """Provide lazy initialization for the underlying httpx.AsyncClient.
 
         Returns:
